call-debit-spread-finder/pyproject.toml
[tool.poetry]
name = "call-debit-spread-finder"
version = "0.1.0"
description = "Find golden Call Debit Spread opportunities from unusual options signals"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "cds_finder", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
rich = "^13.7.0"
supabase = "^2.7.0"
yfinance = "^0.2.32"
pandas = "^2.1.0"
numpy = "^1.26.0"
scipy = "^1.11.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
ruff = "^0.1.0"

[tool.poetry.scripts]
cds-finder = "cds_finder.cli:app"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

